+++ src/action/evaluate.py
@@ -1,16 +1,152 @@
 # -*- coding: utf-8 -*-
+"""
+"""
 from __future__ import unicode_literals, print_function
+
+import re
+import string
 
 from django.core.exceptions import ObjectDoesNotExist
 from django.template import Context, Template, TemplateSyntaxError
 from django.template.loader import render_to_string
 from validate_email import validate_email
+from django.utils.html import escape
 
 import dataops.formula_evaluation
 from action.models import Condition
 from dataops import pandas_db, ops
 from ontask import OntaskException
 from workflow.models import Workflow
+
+def make_xlat(*args, **kwds):
+    """
+    Auxuliary function to define a translator that applies multiple character
+    substitutions at once.
+
+    Taken from "Python Cookbook, 2nd Edition by David Ascher, Anna Ravenscroft,
+    Alex Martelli", Section 1.18
+
+    :param args: Dictionary
+    :param kwds:
+    :return: A function that uses the given dictionary to apply multiple
+    changes to a string
+    """
+    adict = dict(*args, **kwds)
+    rx = re.compile('|'.join(map(re.escape, adict)))
+
+    def one_xlat(match):
+        return adict[match.group(0)]
+
+    def xlat(text):
+        return rx.sub(one_xlat, text)
+
+    return xlat
+
+
+# Dictionary to translate non alphanumeric symbols into alphanumeric pairs
+# 0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ
+# !"#$%&'()*+,-./:;<=>?@[\]^_`{|}~
+# abcdefghijklmnopqrstuvwxyzABCDEFG
+tr_item = make_xlat({
+    '!': '_a', '"': '_b', '#': '_c', '$': '_d', '%': '_e', '&': '_f',
+    "'": "_g", '(': '_h', ')': '_i', '*': '_j', '+': '_k', ',': '_l',
+    '-': '_m', '.': '_n', '/': '_o', ':': '_p', ';': '_q', '<': '_r',
+    '=': '_s', '>': '_t', '?': '_u', '@': '_v', '[': '_w', '\\': '_x',
+    ']': '_y', '^': '_z', '_': '_0', '`': '_1', '{': '_2', '|': '_3',
+    '}': '_4', '~': '_5', ' ': '_6',
+})
+
+
+def translate(varname):
+    """
+    Function that given a string representing a variable name applies a
+    translation to each of the non alphanumeric characters in that name.
+    Additionally, it needs to guarantee that the name starts with a letter
+    (not a digit), and it detects and fixes this condition by introducing a
+    prefix.
+
+    :param varname: Variable name
+    :return: New variable name starting with a letter followed only by
+             letter, digit or _
+    """
+    # If the variable name starts with a non-letter or the prefix used to
+    # force letter start, add a prefix.
+    if not varname[0] in string.letters or varname.startswith('OT_'):
+        varname = 'OT_' + varname
+
+    # Return the new variable name surrounded by the detected marks.
+    return tr_item(varname)
+
+
+def render_template(template_text, context_dict):
+    """
+    Given a template text and a context, performs the rendering of the
+    template using the django template mechanism but with an additional
+    pre-processing to bypass the restriction imposed by Jinja/Django that the
+    variable names must start with a letter followed by a letter, number or _.
+
+    In OnTask, the variable names are: column names, attribute names,
+    or condition names. It is too restrictive to propagate the restrictions
+    imposed by Jinja variables all the way to these three components. To
+    shield this from the users, there is a preliminary step in which those
+    variables in the template and keys in the context that do not comply with
+    the syntax restriction are renamed to compliant names.
+
+    The problem:
+    We are giving two objects: a string containing a template with two markup
+    structures denoting the use of variables, and a dictionary that matches
+    variables to values.
+
+    The processing of the template is done through Django template engine
+    (itself based in Jinja). The syntax for the variables appearing in the
+    template is highly restrictive. It only allows names starting with a
+    letter followed by letter, number or '_'. No other printable symbol is
+    allowed. We want to relax this last restriction.
+
+    The solution:
+    1) Parse the template and detect the use of all variables.
+
+    2) For each variable use, transform its name into a name that is legal
+       for Jinja (starts with letter followed by letter, number or '_' *)
+
+       The transformation is based on:
+       - Every non-letter or number is replaced by '_' followed by a
+         letter/number as specified by the dictionary below.
+
+       - If the original variable does not start by a letter, insert a prefix.
+
+    3) Apply the same transformation for the keys in the given dictionary
+
+    4) Execute the new template with the new dictionary and return the result.
+
+    :param template_text: Text in the template to be rendered
+    :param context_dict: Dictionary used by Jinja to evaluate the template
+    :return: The rendered template
+    """
+
+    # Regular expression detecting the use of a variable, or the
+    # presence of a "{% if variable %} construct in a string (template)
+    var_use_re = re.compile(
+        '(?P<markup_pre>{({|% if) )(?P<varname>.+?)(?P<markup_post> [%\}]\})')
+
+    # Steps 1 and 2. Apply the tranlation process to all variables that
+    # appear in the the template text
+    new_template_text = var_use_re.sub(
+        lambda m: m.group('markup_pre') + \
+                  translate(m.group('varname')) + \
+                  m.group('markup_post'),
+        template_text)
+
+    # Step 3. Apply the translation process to the context keys
+    new_context = dict([(translate(escape(x)), y)
+                        for x, y in context_dict.items()])
+
+    # If the number of elements in the two dictionaries is different, we have
+    #  a case of collision in the translation. Need to stop immediately.
+    assert len(context_dict) == len(new_context)
+
+    # Step 4. Return the redering of the new elements
+    return Template(new_template_text).render(Context(new_context))
 
 
 def evaluate_action(action, extra_string, column_name):
@@ -44,18 +180,6 @@ def evaluate_action(action, extra_string, column_name):
     col_idx = -1
     if column_name and column_name in col_names:
         col_idx = col_names.index(column_name)
-
-    # Get the attributes from the workflow
-    attributes = workflow.attributes
-
-    template = Template(action.content)
-    extra_template = None
-    if extra_string:
-        # Create a template for the extra string
-        try:
-            extra_template = Template(extra_string)
-        except Exception as e:
-            return 'Syntax error detected in the subject. ' + e.message
 
     # Step 2: Get the row of data from the DB
     try:
@@ -98,15 +222,22 @@ def evaluate_action(action, extra_string, column_name):
 
         # Step 4: Create the context with the attributes, the evaluation of the
         # conditions and the values of the columns.
+        attributes = workflow.attributes
         context = dict(dict(row_values, **condition_eval), **attributes)
 
         # Step 5: run the template with the given context
         # Render the text and append to result
-        partial_result = [template.render(Context(context))]
+        try:
+            partial_result = [render_template(action.content, context)]
+        except Exception as e:
+            return 'Syntax error detected in the action text. ' + e.message
 
         # If there is extra message, render with context and create tuple
         if extra_string:
-            partial_result.append(extra_template.render(Context(context)))
+            try:
+                partial_result.append(render_template(extra_string, context))
+            except Exception as e:
+                return 'Syntax error detected in the subject. ' + e.message
 
         # If column_name was given (and it exists), create a tuple with that
         # element as the third component
@@ -193,10 +324,66 @@ def evaluate_row(action, row_idx):
     # Step 5: run the template with the given context
     # First create the template with the string stored in the action
     try:
-        template = Template(action.content)
+        result = render_template(action.content, context)
     except TemplateSyntaxError as e:
         return render_to_string('action/syntax_error.html',
                                 {'msg': e.message})
 
     # Render the text
-    return template.render(Context(context))
+    return result
+
+def run(*script_args):
+    """
+    Script for testing purposes
+    :param script_args:
+    :return:
+    """
+
+    template = """
+    hi --{{ one }}--
+    --{% if var1 %}var1{% endif %}--
+    --{% if var2 %}var2{% endif %}-- 
+    --{% if !"# %}var3{% endif %}--
+    --{% if $%& %}var4{% endif %}--
+    --{% if '() %}var5{% endif %}--
+    --{{ +,- }}--
+    --{{ ./: }}--
+    --{{ ;<= }}--
+    --{{ >?@ }}--
+    --{{ [\] }}--
+    --{{ ^_` }}--
+    --{{ {|}~ }}--
+    --{{ this one has spaces }}--
+    --{{ OT_ The prefix }}--
+    --{{ OT_The prefix 2 }}--
+    """
+
+    template = u'<p>Hi&nbsp;{{ !"#$%&amp;()*+,-./:;&lt;=&gt;?@[\\]^_`{|}~ }}</p>'
+
+    context = {
+        'one': 1,
+        'var1': True,
+        'var2': True,
+        '!"#': True,
+        '$%&': True,
+        "'()": True,
+        '+,-': 'var6',
+        './:': 'var7',
+        ';<=': 'var8',
+        '>?@': 'var9',
+        '[\]': 'var10',
+        '^_`': 'var11',
+        '{|}~': 'var12',
+        'this one has spaces': 'The spaces are not a problem',
+        'OT_ The prefix': 'Prefix solved.',
+        'OT_The prefix 2': 'Prefix 2 solved',
+        'The prefix 2': 'This should NOT appear. ERROR',
+    }
+    context = {
+        u'!"#$%&()*+,-./:;<=>?@[\\]^_`{|}~': u'Carmelo Coton',
+    }
+
+    print(escape(context.items()[0][0]))
+    print(render_template(template, context))
+
+