+++ src/action/models.py
@@ -1,13 +1,19 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals, print_function
 
+import re
+
 import datetime
 import pytz
 from django.conf import settings
 from django.contrib.postgres.fields import JSONField
 from django.db import models
+from django.utils.html import escape
 
 from workflow.models import Workflow, Column
+
+# Regular expression to detect the use of a variable in a django template
+var_use_re = re.compile('{{ (?P<varname>.+?) \}\}')
 
 
 class Action(models.Model):
@@ -101,7 +107,28 @@ class Action(models.Model):
         return not ((self.active_from and now < self.active_from) or
                     (self.active_to and self.active_to < now))
 
+    def rename_variable(self, old_name, new_name):
+        """
+        Function that renames a variable present in the action content
+        :param old_name: Old name of the variable
+        :param new_name: New name of the variable
+        :return: Updates the current object
+        """
+
+        new_text = var_use_re.sub(
+            lambda m: '{{ ' +
+                      (new_name if m.group('varname') == escape(old_name)
+                       else m.group('varname')) + ' }}',
+            self.content
+        )
+        self.content = new_text
+        self.save()
+
     class Meta:
+        """
+        Define the criteria of uniqueness with name in workflow and order by
+        name
+        """
         unique_together = ('name', 'workflow')
         ordering = ('name',)
 
@@ -137,5 +164,10 @@ class Condition(models.Model):
         return self.name
 
     class Meta:
+        """
+        The unique criteria here is within the action, the name and being a
+        filter. We may choose to name a filter and a condition with the same
+        name (no need to restrict it)
+        """
         unique_together = ('action', 'name', 'is_filter')
         ordering = ('created',)