+++ src/table/views.py
@@ -8,11 +8,11 @@ from django.contrib.auth.decorators import user_passes_test
 from django.http import JsonResponse
 from django.shortcuts import redirect, reverse, render
 from django.template.loader import render_to_string
+from django.utils.html import escape
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_http_methods
 
-from dataops import ops
-from dataops import pandas_db
+from dataops import ops, pandas_db
 from ontask.permissions import is_instructor
 from workflow.ops import get_workflow
 
@@ -53,7 +53,7 @@ def display_ss(request):
         draw = int(request.POST.get('draw', None))
         start = int(request.POST.get('start', None))
         length = int(request.POST.get('length', None))
-        order_col = request.POST.get('order[0][column]', None)
+        order_col_name = request.POST.get('order[0][column]', None)
         order_dir = request.POST.get('order[0][dir]', 'asc')
     except ValueError:
         return JsonResponse({'error': 'Incorrect request. Unable to process'})
@@ -66,8 +66,9 @@ def display_ss(request):
     column_names = [x.name for x in columns]
 
     # See if an order has been given.
-    if order_col:
-        order_col = column_names[int(order_col) - 1]  # The first column is ops
+    if order_col_name:
+        # The first column is ops
+        order_col_name = column_names[int(order_col_name) - 1]
 
     # Find the first key column
     key_name, key_idx = next(((c.name, idx) for idx, c in enumerate(columns)
@@ -82,7 +83,7 @@ def display_ss(request):
         workflow.id,
         cv_tuples,
         True,
-        order_col,
+        order_col_name,
         order_dir == 'asc',
         None
     )
@@ -95,15 +96,15 @@ def display_ss(request):
         ops_string = render_to_string(
             'table/includes/partial_table_ops.html',
             {'edit_url':
-             reverse('dataops:rowupdate') +
-             '?update_key={0}&update_val={1}'.format(key_name,
-                                                     row[key_idx]),
+                 reverse('dataops:rowupdate') +
+                 '?update_key={0}&update_val={1}'.format(key_name,
+                                                         row[key_idx]),
              'delete_key': '?key={0}&value={1}'.format(key_name, row[key_idx]),
              }
         )
-        final_qs.append(OrderedDict(
-            [('Ops', ops_string)] + zip(column_names, row))
-        )
+        # Create the list of elements to display and add it ot the final QS
+        final_qs.append([ops_string] + list(row))
+
         if items == length:
             # We reached the number or requested elements, abandon.
             break